feat: Now create one stitched gif from two downloaded gifs
The gif for looking at the time/watch is now a random downloaded gif from tenor and gets stitched together with another random gif from tenor that matches the specified query (e.g. Overwatch -> a random gif about Overwatch is downloaded) and has the query text added with "time for" in front of it.